Zobrist hashing is done

+++ src/gamestate/defs.rs
@@ -1,5 +1,5 @@
 pub const PIECE_TYPES_NUM: usize = 6;
-pub const BOARD_SIDE_LENGTH: u8 = 8;
+pub const BOARD_SIDE_LENGTH: usize = 8;
 pub const BOARD_NUM_OF_SQUARES: usize = 64;
 pub const SIDE_NUM: usize = 2;
 pub const CASTLING_CONFIGURATIONS_NUM: usize = 16;
